select

+++ ephub-common-rpc-connector/src/main/java/com/shangpin/ephub/client/data/mysql/hub/dto/HubWaitSelectRequestDto.java
@@ -20,9 +20,9 @@ public class HubWaitSelectRequestDto {
     private String spuModel;
     private String brandNo;
     private String categoryNo;
-    private int spuState;
+    private Byte spuState;
     private String startDate;
     private String endDate;
-    private int spuSelectState;
+    private Byte spuSelectState;
 	  
 }
