+++ ephub-common-rpc-connector/src/main/java/com/shangpin/ephub/client/data/mysql/hub/dto/HubWaitSelectRequestWithPageDto.java
@@ -22,10 +22,10 @@ public class HubWaitSelectRequestWithPageDto {
     private String spuModel;
     private String brandNo;
     private String categoryNo;
-    private int spuState;
+    private Byte spuState;
     private String startDate;
     private String endDate;
-    private int spuSelectState;
+    private Byte spuSelectState;
     
     public Integer getPageNo() {
         return pageNo;
